requirements.txt
# Dependencies
boto3>=1.34.0
botocore>=1.34.0
python-jenkins>=1.8.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
typer[all]>=0.9.0  # CLI interface with rich output
sqlalchemy>=2.0.0  # Database ORM
alembic>=1.13.0  # Database migrations
asyncpg>=0.29.0  # PostgreSQL async driver
aiosqlite>=0.19.0  # SQLite async driver
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiohttp>=3.9.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
types-boto3>=1.0.2
types-python-dateutil>=2.8.19
